requirements.txt
# Dependency versi utama
pandas
numpy
Pillow

# Development dependencies
python-dotenv
black
pytest
scikit-learn